Add a Small logical volume to gzip backup script
This script is meant to be launched from you cloud-manager server. It
creates a snapshot of every LVM volumes, gzip it, and sent it to a
directory.
The difference between the
SCR_5004_$version_NUAC-Sauvegarde-Volumes-EBS.sh script is that the
logical volume is backed up to a .img.gz file. It is meant to be used
in order to have
mages backup that could be easily reinjected into another logical volume